Bug Fixes and (Almost Done) Code Reorganization
Reorganization of Code (Because the daisy chains of functions was dumb, just set it all up from a main thread to keep track of it). This will make it easier to track down bugs, so I can make sure when I move over to an async task, the only bugs there are from me screwing that up, not the code beforehand.

Big fixes to percent calculation were added because I fixed a mistake where I mixed up the 2d array position 1 and 3, or in the passing time case, didn't subtract one so that it was the end time before passing time.

Get hyped for version 1.04.