Final Version 1.04: Calculation, Trimming of old code, Fix to offline mode
Fixed Calculation of Time with 0% left, so that it switches to the period of time going forward.

Lots of old code sitting around was taken out of the file, and additional documentation was added.

Offline mode was broken in the reorganization, but was fixed by making sure the code waited for the user to respond to a popup.

This will be the code going into the final 1.04 update. Next update will have no refresh button, and a redesigned period display (I think) that shows the actual period times.