Beginning of UI Overhaul
In preparation for showing of schedules with class names and times. I'm sorry Seth, the circle was way too big. Even though the customer might have been right, I refuse to accept it.

Maybe I'll add a legacy UI for all those who desire to stay with version 1.6.

I'm not an iStaples clone, I swear. I have the hamburger menu in the top left corner. They basically copied myStaples which was around before iStaples, so I'm really just copying a copy of a copy of an app. That is, if I'm actually copying. Which I'm not. Totally not.

Don't look at me like that! They don't own the rights to a circular progress bar smaller than half the screen!