Fixed Major Array Misalignment Bug
App now does not crash randomly because one ArrayList is bigger than the other, or try to read a null array.

Code also looks a lot nicer because I took out lines I was repeating code for no reason in.

Noted an issue with FindTimeUntilEndPassingTime(lunchWaveTimes, tempLunchPeriod), since it can't distinguish before the lunch wave occurs but after period 3. I think I might have to consolidate the code for the calculation with the display functions in order to save processing power and account for small conditions like this one.