Almost Completed, needs to account for one more case though
Basically everything highlights and calculates when it's supposed to unless its during a lunch wave passing time where the length of a period jumps between lunch waves. Passing Time is Triggered there, and for 5 minutes it will treat the class as if it had a passing time during that time, then switch back to normal. Will fix later.