Full Refresh Capabilities and Addition of Offline Snackbar Warning
App actually refreshes when it detects the time has changed, not something dumb like every 15 seconds. Added a warning so people that are using an offline database know that they're using it, and when it was updated. Decided not to code anything for first time users without an internet connection, decided to simply insert a dialog telling them to connect to the internet before using offline mode.